backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
numpy>=1.26.0
Pillow>=10.0.0
scipy>=1.11.0
pydantic>=2.5.0
python-multipart>=0.0.6
